feat: ToCoursewarePage gets its own component
